// src/types.ts
export type TCarInfo = {
  modele: string
  finitions: { version: string[] }
  carrosserie: string
  genre_v: string
  hauteur: string
  longueur: string
  marque: string
  marque_carros: string
  nb_pl_ass: any
  version: any
  prix_vehic: string
  carrosserie_cg: string
  genre_v_cg: any
  largeur: string
  modele_etude: string
  modele_prf: string
  nb_portes: string
  type_var_vers: any
  co2: string
  cylindree: string
  empat: string
  depollution: string
  energie: string
  mode_inject: string
  nb_cylind: string
  nb_vitesses: any
  turbo_compr: string
  code_moteur: string
  montes_pneu_compatible: any
  cons_urb: string
  poids_vide: string
  nb_volumes: any
  propulsion: string
  ptr: string
  ptr_prf: string
  puis_ch: string
  puis_fisc: string
  tb_boite_vit: any
  cons_mixte: string
  puis_kw: string
  cons_exurb: string
  codif_vin_prf: string
  couleur_vehic: string
  immat_siv: string
  n_serie: string
  type: string
  type_vin_cg: string
  siren_ent: any
  date_1er_cir: string
  date_dcg: string
}